Recommit after deletion and change filepath stored
This file has been recommited so that it can be in a folder
under the new arrangement.
The contents have also been slightly changed, so that it does not store the
filepath of the current working directory. Instead it stores the path of the folder
containing the working directory (where the working directory is a sub folder).